# Make parser "safe"