… benchmarks/Kestrel.Performance/Http1ConnectionBenchmark.cs
… benchmarks/Kestrel.Performance/Http1ConnectionParsingOverheadBenchmark.cs
… benchmarks/Kestrel.Performance/HttpParserBenchmark.cs
… benchmarks/Kestrel.Performance/KnownStringsBenchmark.cs
… benchmarks/Kestrel.Performance/Mocks/NullParser.cs
… benchmarks/Kestrel.Performance/RequestParsingBenchmark.cs
… src/Kestrel.Core/Internal/Http/Http1Connection.cs
… src/Kestrel.Core/Internal/Http/Http1ParsingHandler.cs
… src/Kestrel.Core/Internal/Http/HttpParser.cs
… src/Kestrel.Core/Internal/Http/HttpProtocol.cs
… src/Kestrel.Core/Internal/Http/HttpRequestHeaders.cs
… src/Kestrel.Core/Internal/Http/IHttpHeadersHandler.cs
… src/Kestrel.Core/Internal/Http/IHttpParser.cs
… src/Kestrel.Core/Internal/Http/IHttpRequestLineHandler.cs
… src/Kestrel.Core/Internal/Http/PathNormalizer.cs
… src/Kestrel.Core/Internal/Http2/Http2Connection.cs
… src/Kestrel.Core/Internal/Infrastructure/HttpCharacters.cs
… src/Kestrel.Core/Internal/Infrastructure/HttpUtilities.cs
… src/shared/ThrowHelper.cs
… test/Kestrel.Core.Tests/AsciiDecoding.cs
… test/Kestrel.Core.Tests/BufferReaderTests.cs
… test/Kestrel.Core.Tests/HPackDecoderTests.cs
… test/Kestrel.Core.Tests/Http1ConnectionTests.cs
… test/Kestrel.Core.Tests/HttpParserTests.cs
… test/Kestrel.Core.Tests/HttpUtilitiesTest.cs
… test/Kestrel.Core.Tests/KnownStringsTests.cs
… test/Kestrel.InMemory.FunctionalTests/BadHttpRequestTests.cs
… test/Kestrel.InMemory.FunctionalTests/Http2/Http2TestBase.cs
… test/shared/HttpParsingData.cs