@JeremyKuhne 2018-10-31
Added the core of the BufferReader struct changes. Working on the actual meat of the parser changes now.