@JeremyKuhne 2018-11-01
Fyi, I'm using this as a sanity check for getting `SequenceReader` into CoreFX.